% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_map.R
\name{g_map}
\alias{g_map}
\title{Local G Map}
\usage{
g_map(
  polys,
  xname,
  weights = NULL,
  permutations = 999,
  alpha = 0.05,
  type = "g"
)
}
\arguments{
\item{polys}{An sf dataframe}

\item{xname}{string, the name of the x variable, this variable must be contained in the sf dataframe}

\item{weights}{weights structure from spdep, must be style "B"; default is set equal to NULL, and first
order queen contiguity weights are used to construct the map}

\item{permutations}{numeric, number of permutations the conditional randimization approach to significance, maximum is 99999,
default is 999}

\item{alpha}{numeric, cut-off level of significance, must be between 0 and 1, the default is .05}

\item{type}{string, can be "g" or "gstar"}
}
\description{
The function to make local G and G* cluster maps.
}
