% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{local_g}
\alias{local_g}
\title{Local G Statistics}
\usage{
local_g(x, weights, type = "g")
}
\arguments{
\item{x}{A vector of numerical values}

\item{weights}{Weights structure from spdep, must be style "B"}

\item{type}{A string, designates g or gstar}
}
\value{
local A vector of local geary statistics
}
\description{
The function to compute G and G* statistics
}
