% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{local_g_sims}
\alias{local_g_sims}
\title{Local G Simulations}
\usage{
local_g_sims(x, weights, permutations, type = "g")
}
\arguments{
\item{x}{A vector of numerical values}

\item{weights}{Weights structure from spdep, must be style "B"}

\item{permutations}{Number of permutations, the default is 999}

\item{type}{designates the type of statistic g or gstar, g is the default}
}
\value{
local.sims Reference distributions of local geary statistics for each location in matrix form
}
\description{
Function to compute reference distributions of local G and G* statistics
}
