% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geary.R
\name{local_geary}
\alias{local_geary}
\title{Local Geary Statistics}
\usage{
local_geary(x, y = NULL, weights)
}
\arguments{
\item{x}{A vector of numerical values}

\item{y}{A vector of numerical values, only to be used in the bivariate case}

\item{weights}{Weights structure from spdep, must be style "B"}
}
\value{
local A vector of local geary statistics
}
\description{
The function to compute univariate and bivariate local geary statistics
}
