% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geary.R
\name{local_geary_pvalue}
\alias{local_geary_pvalue}
\title{Local Geary P-values}
\usage{
local_geary_pvalue(x, y = NULL, weights, permutations = 999)
}
\arguments{
\item{x}{A vector of numerical values}

\item{y}{A vector of numerical values, only to be used in the bivariate case}

\item{weights}{Weights structure from spdep, must be style "B"}

\item{permutations}{Number of permutations, the default is 999}
}
\value{
pvalue A vector of p-values for each locations local geary statistic
}
\description{
Function to compute the p-value of the observed local geary statistics under the
conditional randomizaton approach
}
