% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joincount.R
\name{local_jc}
\alias{local_jc}
\title{Local Join Count Statistics}
\usage{
local_jc(x, y = NULL, weights)
}
\arguments{
\item{x}{A vector of numerical values}

\item{y}{A vector of numerical values, default is NULL}

\item{weights}{Weights structure from spdep, must be style "B"}
}
\value{
local A vector of local joint count statistics
}
\description{
The function to compute univariate and bivariate local join count statistics
}
