% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joincount.R
\name{local_jc_sims}
\alias{local_jc_sims}
\title{Local Join Count Simulations}
\usage{
local_jc_sims(x, y = NULL, weights, permutations)
}
\arguments{
\item{x}{A vector of numerical values}

\item{weights}{Weights structure from spdep, must be style "B"}

\item{permutations}{Number of permutations, the default is 999}

\item{z}{A vector of numerical values, only to be used in the bivariate case, default is NULL}
}
\value{
local.sims Reference distributions of local join count statistics for each location in matrix form
}
\description{
Function to compute reference distributions of local join count statistics
}
