% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{local_moran}
\alias{local_moran}
\title{Local Moran Statistics}
\usage{
local_moran(x, y = NULL, weights)
}
\arguments{
\item{x}{A vector of numerical values}

\item{y}{A vector of numerical values, only to be used in the bivariate case}

\item{weights}{Weights structure from spdep, must be style "B"}
}
\value{
local A vector of local moran statistics
}
\description{
The function to compute univariate and bivariate local moran statistics
}
