% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{local_moran_sims}
\alias{local_moran_sims}
\title{Local Moran Simulations}
\usage{
local_moran_sims(x, y = NULL, weights, permutations = 999)
}
\arguments{
\item{x}{A vector of numerical values}

\item{y}{A vector of numerical values, only to be used in the bivariate case}

\item{weights}{Weights structure from spdep, must be style "B"}

\item{permutations}{Number of permutations, the default is 999}
}
\value{
local.sims Reference distributions of local moran statistics for each location in matrix form
}
\description{
Function to compute reference distributions of local moran statistics
}
