% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_map.R
\name{moran_map}
\alias{moran_map}
\title{Local Moran Map}
\usage{
moran_map(
  polys,
  xname,
  yname = NULL,
  weights = NULL,
  permutations = 999,
  alpha = 0.05
)
}
\arguments{
\item{polys}{An sf dataframe}

\item{xname}{string, the name of the x variable, this variable must be contained in the sf dataframe}

\item{yname}{string, the name of the y variable, this variable must be contained in the sf dataframe, default option
is NULL, can only be used for moran, geary, and join count}

\item{weights}{weights structure from spdep, must be style "B"; default is set equal to NULL, and first
order queen contiguity weights are used to construct the map}

\item{permutations}{numeric, number of permutations the conditional randimization approach to significance, maximum is 99999,
default is 999}

\item{alpha}{numeric, cut-off level of significance, must be between 0 and 1, the default is .05}
}
\description{
The function to make local moran cluster maps.
}
