% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_map.R
\name{significance_map}
\alias{significance_map}
\title{Significance Map}
\usage{
significance_map(
  polys,
  xname,
  yname = NULL,
  type,
  weights = NULL,
  alpha = 0.05,
  permutations = 999
)
}
\arguments{
\item{polys}{An sf dataframe}

\item{xname}{string, the name of the x variable, this variable must be contained in the sf dataframe}

\item{yname}{string, the name of the y variable, this variable must be contained in the sf dataframe, default option
is NULL, can only be used for moran, geary, and join count}

\item{type}{string, the type of local statistic, options are: "moran", "geary", "g", "gstar", and "join_count"}

\item{weights}{weights structure from spdep, must be style "B"; default is set equal to NULL, and first
order queen contiguity weights are used to construct the map}

\item{alpha}{numeric, cut level of significance, must be between 0 and 1, the default is .05}

\item{permutations}{numeric, number of permutations the conditional randimization approach to significance, maximum is 99999,
default is 999}
}
\description{
The function to make significance maps for a variety of local statistics. These
statistics include moran, geary, G, G*, and join count. There are bivariate options for moran,
geary, and join count
}
